OData class creation
