+++ msticpy/data/drivers/mdatp_driver.py
@@ -4,43 +4,20 @@
 # license information.
 # --------------------------------------------------------------------------
 """MDATP OData Driver class."""
-from typing import Tuple, Union, Any, Dict
-import re
-import urllib
-
-import requests
-
+from typing import Union, Any
 import pandas as pd
 
-from .driver_base import DriverBase
+from .odata_driver import OData
 from ...nbtools.utility import export
-from ...nbtools import pkg_config as config
 from ..._version import VERSION
 
 __version__ = VERSION
 __author__ = "Pete Bryan"
 
-_OAUTH_URL = "https://login.windows.net/{tenantId}/oauth2/token"
-
 
 @export
-class MDATPDriver(DriverBase):
+class MDATPDriver(OData):
     """KqlDriver class to retreive date from MDATP."""
-
-    _REQ_BODY = {
-        "client_id": None,
-        "client_secret": None,
-        "grant_type": "client_credentials",
-        "resource": "https://api.securitycenter.windows.com",
-    }
-    _REQ_HEADER = {
-        "Content-Type": "application/json",
-        "Accept": "application/json",
-        "Authorization": None,
-    }
-
-    _DEF_API_ROOT = "https://api.securitycenter.windows.com/"
-    _DEF_API_VER = "api"
 
     def __init__(self, connection_str: str = None, **kwargs):
         """
@@ -52,68 +29,16 @@ class MDATPDriver(DriverBase):
             Connection string
 
         """
-        super().__init__()
-        self._loaded = True
-        self.aad_token = None
-        self.req_headers = None
-        self.api_root = ""
-        self._debug = kwargs.get("debug", False)
-
-        if connection_str:
-            self.current_connection = connection_str
-            self.connect(connection_str)
-
-    def connect(self, connection_str: str = None, **kwargs):
-        """
-        Connect to MDATP data source.
-
-        Parameters
-        ----------
-        connection_string : str
-            Connect to a data source
-
-        Notes
-        -----
-        Connection string fields:
-        tenantId
-        clientId
-        clientSecret
-        apiRoot
-        apiVersion
-
-        """
-        if connection_str:
-            self.current_connection = connection_str
-            cs_dict = self._parse_connection_str(connection_str)
-        elif kwargs:
-            cs_dict = kwargs
-        else:
-            cs_dict = config.settings.get("MDATPApp")["Args"]
-
-        req_url = _OAUTH_URL.format(tenantId=cs_dict["tenantId"])
-        req_body = dict(self._REQ_BODY)
-        req_body["client_id"] = cs_dict["clientId"]
-        req_body["client_secret"] = cs_dict["clientSecret"]
-
-        # authenticate and obtain AAD Token for future calls
-        data = urllib.parse.urlencode(req_body).encode("utf-8")
-        response = requests.post(url=req_url, data=data)
-        json_response = response.json()
-        self.aad_token = json_response.get("access_token", None)
-        if not self.aad_token:
-            raise ConnectionError("Could not obtain access token")
-
-        self.req_headers = dict(self._REQ_HEADER)
-        self.req_headers["Authorization"] = "Bearer " + self.aad_token
-        self.api_root = cs_dict.get("apiRoot", self._DEF_API_ROOT) + cs_dict.get(
-            "apiVersion", self._DEF_API_VER
-        )
-
-        print("Connected.")
-        self._connected = True
-
-        json_response["access_token"] = None
-        return json_response
+        super().__init__(connection_str)
+        self.req_body = {
+            "client_id": None,
+            "client_secret": None,
+            "grant_type": "client_credentials",
+            "resource": "https://api.securitycenter.windows.com",
+        }
+        self.oauth_url = "https://login.windows.net/{tenantId}/oauth2/token"
+        self.api_root = "https://api.securitycenter.windows.com/"
+        self.api_ver = "api"
 
     def query(self, query: str) -> Union[pd.DataFrame, Any]:
         """
@@ -131,110 +56,6 @@ class MDATPDriver(DriverBase):
             Kql ResultSet if an error.
 
         """
-        return self.query_with_results(query)[0]
-
-    # pylint: disable=too-many-branches
-    def query_with_results(self, query: str) -> Tuple[pd.DataFrame, Any]:
-        """
-        Execute query string and return DataFrame of results.
-
-        Parameters
-        ----------
-        query : str
-            The kql query to execute
-
-        Returns
-        -------
-        Tuple[pd.DataFrame, results.ResultSet]
-            A DataFrame (if successfull) and
-            Kql ResultSet.
-
-        """
-        if not self.connected:
-            self.connect(self.current_connection)
-            if not self.connected:
-                raise ConnectionError(
-                    "Source is not connected. ", "Please call connect() and retry."
-                )
-
-        if self._debug:
-            print(query)
-
-        req_url = self.api_root + "/advancedqueries/run"
-        req_url = urllib.parse.quote(req_url, safe="%/:=&?~#+!$,;'@()*[]")
-        body = {"Query": query}
-        response = requests.post(url=req_url, headers=self.req_headers, data=str(body))
-        if response.status_code != requests.codes["ok"]:
-            if response.status_code == 401:
-                raise ConnectionRefusedError(
-                    "Authentication failed - possible ", "timeout. Please re-connect."
-                )
-            # Raise an exception to handle hittng the Advanced Hunting API limits
-            if response.status_code == 429:
-                raise ConnectionRefusedError(
-                    "You have likely hit the MDATP API limit, please wait some time and try again. https://docs.microsoft.com/en-us/windows/security/threat-protection/microsoft-defender-atp/run-advanced-query-api"  # pylint: disable=line-too-long
-                )
-            response.raise_for_status()
-
-        json_response = response.json()
-        if isinstance(json_response, int):
-            print(
-                "Warning - query did not complete successfully.",
-                "Check returned response.",
-            )
-            return None, json_response
-
-        if "Results" in json_response:
-            result = json_response["Results"]
-        else:
-            result = json_response
-
-        if not result:
-            print("Warning - query did not return any results.")
-            return None, json_response
-        return pd.io.json.json_normalize(result), result
-
-    # pylint: enable=too-many-branches
-
-    @staticmethod
-    def _parse_connection_str(connection_str: str) -> Dict[str, str]:
-        """
-        Split connection string components into dictionary.
-
-        Parameters
-        ----------
-        connection_str : str
-            Semi-colon delimited connection string
-
-        Returns
-        -------
-        Dict[str, str]
-            dict of key/pair values
-
-        """
-        cs_items = connection_str.split(";")
-        cs_dict = {
-            prop[0]: prop[1]
-            for prop in [item.strip().split("=") for item in cs_items]
-            if prop[0] and prop[1]
-        }
-        return cs_dict
-
-    @staticmethod
-    def _prepare_param_dict_from_filter(filterstr: str) -> Dict[str, str]:
-        """
-        Parse filter string into dictionary.
-
-        Parameters
-        ----------
-        filterstr : str
-            OData filter string
-
-        """
-        get_params = {}
-        for filter_param in re.split(r"[\?\&]+", filterstr):
-            if filter_param:
-                attr = filter_param.split("=")[0]
-                val = filter_param.split("=")[1]
-                get_params[attr] = val
-        return get_params
+        return self.query_with_results(
+            query, body=True, api_end="/advancedqueries/run"
+        )[0]
